misc/HtmlTestSuite/results/test-0503.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Grid: width of grid container with scrollbar.</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas"/>
<link href="https://drafts.csswg.org/css-grid-1/#overflow" rel="help"/>
<meta content="This test ensures that a grid container scrollbar is computed properly during intrinsic width calculation." name="assert"/>

<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="/css/support/width-keyword-classes.css" rel="stylesheet"/>
<style>
  .grid {
    overflow-y: scroll;
    grid-template-columns: repeat(4, 50px);
  }
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
</head><body onload="checkLayout('.grid')">
  <div class="grid min-content" data-expected-client-width="200">
    item
  </div>
  <div class="grid max-content" data-expected-client-width="200">
    item
  </div>
</body>
</html>
